env + защита эндпоинтов + пара фич

+++ app/routers/posts.py
@@ -1,7 +1,7 @@
 from fastapi import status, HTTPException, Depends, APIRouter
 from sqlalchemy.orm import Session
 from ..database import get_db
-from .. import models, schemas
+from .. import models, schemas, oauth2
 import typing
 
 router = APIRouter(
@@ -10,8 +10,11 @@ router = APIRouter(
     )
 
 @router.get("/", response_model=typing.List[schemas.PostResponse])
-def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
+def get_posts(db: Session = Depends(get_db), limit: int = 5, page: int = 1, search: str = ""):
+    posts_to_skip = limit * (page - 1)
+
+    posts = db.query(models.Post).where(models.Post.post_title.contains(search.lower())) \
+    .limit(limit=limit).offset(offset=posts_to_skip).all()
     return posts
 
 @router.get("/{id}", response_model=schemas.PostResponse)
@@ -24,8 +27,8 @@ def get_post(id: int, db: Session = Depends(get_db)):
     return post
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
-def create_post(post: schemas.PostCreate, db: Session = Depends(get_db)):
-    new_post = models.Post(**post.model_dump())
+def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
+    new_post = models.Post(user_id=current_user.user_id, **post.model_dump())
 
     db.add(new_post)
     db.commit()
@@ -34,22 +37,29 @@ def create_post(post: schemas.PostCreate, db: Session = Depends(get_db)):
     return new_post
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, db: Session = Depends(get_db)):
+def delete_post(id: int, db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     deleted_post = db.query(models.Post).where(models.Post.post_id == id).first()
 
     if deleted_post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id {id} was not found")
 
+    if deleted_post.user_id != current_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"not authorized to perform requested action")
+
     db.delete(deleted_post, synchronize_session=False)
     db.commit()
 
 @router.put("/{id}", response_model=schemas.PostResponse)
-def update_post(id: int, post: schemas.PostUpdate, db: Session = Depends(get_db)):
+def update_post(id: int, post: schemas.PostUpdate, db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).where(models.Post.post_id == id)
+    finded_post = post_query.first()
 
-    if post_query.first() is None:
+    if finded_post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id {id} was not found")
 
+    if finded_post.user_id != current_user.user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"not authorized to perform requested action")
+    
     post_query.update(post.model_dump(), synchronize_session=False)
     db.commit()
